feat: 헤더 컴포넌트 구현 및 댓글 페이지에 통합
